fix: 修复vant的样式冲突问题
