+++ lib/client/src/components/render/mobile/edit-area/edit-area.js
@@ -12,6 +12,7 @@ import previewSwitch from '../preview-switch/preview-switch'
 /* eslint import/no-webpack-loader-syntax: off */
 import vanStyle from '!!raw-loader!vant/lib/index.css'
 import vant3Style from '!!raw-loader!shared/page-code/style/vant3.style.css'
+import { injectCss, removeVantStyle } from 'shared/util'
 
 export default {
     setup (props) {
@@ -21,43 +22,6 @@ export default {
         const isShowModeList = ref(false)
 
         onMounted(() => {
-            function removeVantStyle () {
-                // 获取所有的样式表
-                const styleSheets = document.styleSheets
-
-                // 遍历样式表
-                for (let i = 0; i < styleSheets.length; i++) {
-                    const styleSheet = styleSheets[i]
-                    // 检查是否为CSSStyleSheet
-                    if (styleSheet instanceof CSSStyleSheet) {
-                        const rules = styleSheet.cssRules || styleSheet.rules
-                        // 遍历样式规则
-                        for (let j = 0; j < rules.length; j++) {
-                            const rule = rules[j]
-                            // 检查是否为CSSStyleRule，并且类名以'.van-'开头
-                            if (rule instanceof CSSStyleRule && rule.selectorText.startsWith('.van-')) {
-                            // 删除该样式规则
-                                styleSheet.deleteRule(j)
-                                j-- // 由于删除了一个规则，需要减小索引
-                            }
-                        }
-                    }
-                }
-            }
-
-            // This function injects a dynamic CSS stylesheet into the document head.
-            function injectCss (styles) {
-                // Create a new <style> element
-                const styleSheet = document.createElement('style')
-                // Set the type attribute of the <style> element to 'text/css'
-                styleSheet.type = 'text/css'
-                // Set the id attribute of the <style> element to 'dynamicstylesheet'
-                styleSheet.setAttribute('id', 'dynamicstylesheet')
-                // Set the innerText of the <style> element to the provided styles
-                styleSheet.innerText = styles
-                // Append the <style> element to the document head
-                document.head.appendChild(styleSheet)
-            }
             removeVantStyle()
             injectCss(framework === 'vue3' ? vant3Style : vanStyle)
         })
