+++ lib/client/src/preview/children/app/app.js
@@ -1,5 +1,9 @@
 import './app.css'
-import { h } from 'bk-lesscode-render'
+import { h, framework } from 'bk-lesscode-render'
+import { injectCss, removeVantStyle } from 'shared/util'
+/* eslint import/no-webpack-loader-syntax: off */
+import vanStyle from '!!raw-loader!vant/lib/index.css'
+import vant3Style from '!!raw-loader!shared/page-code/style/vant3.style.css'
 
 export default {
     name: 'app',
@@ -10,6 +14,10 @@ export default {
     },
     created () {
         this.getUserInfo()
+    },
+    mounted () {
+        removeVantStyle()
+        injectCss(framework === 'vue3' ? vant3Style : vanStyle)
     },
     methods: {
         async getUserInfo () {
