+++ lib/shared/util.js
@@ -337,3 +337,49 @@ export const sharedI18n = () => {
         return global.i18n
     }
 }
+/**
+ * 从文档中删除所有Vant样式，通过删除相应的CSS规则。
+ *
+ * @return {undefined} 此函数不返回值。
+ */
+export function removeVantStyle () {
+    // 获取所有的样式表
+    const styleSheets = document.styleSheets
+
+    // 遍历样式表
+    for (let i = 0; i < styleSheets.length; i++) {
+        const styleSheet = styleSheets[i]
+        // 检查是否为CSSStyleSheet
+        if (styleSheet instanceof CSSStyleSheet) {
+            const rules = styleSheet.cssRules || styleSheet.rules
+            // 遍历样式规则
+            for (let j = 0; j < rules.length; j++) {
+                const rule = rules[j]
+                // 检查是否为CSSStyleRule，并且类名以'.van-'开头
+                if (rule instanceof CSSStyleRule && rule.selectorText.startsWith('.van-')) {
+                    styleSheet.deleteRule(j) // 删除该样式规则
+                    j-- // 由于删除了一个规则，需要减小索引
+                }
+            }
+        }
+    }
+}
+
+/**
+ * 将 CSS 样式注入到文档头部。
+ *
+ * @param {string} styles - 要注入的 CSS 样式。
+ * @return {void} 此函数没有返回值。
+ */
+export function injectCss (styles) {
+    // Create a new <style> element
+    const styleSheet = document.createElement('style')
+    // Set the type attribute of the <style> element to 'text/css'
+    styleSheet.type = 'text/css'
+    // Set the id attribute of the <style> element to 'dynamicstylesheet'
+    styleSheet.setAttribute('id', 'dynamicstylesheet')
+    // Set the innerText of the <style> element to the provided styles
+    styleSheet.innerText = styles
+    // Append the <style> element to the document head
+    document.head.appendChild(styleSheet)
+}
